requirements.txt
streamlit
firecrawl-py
ipython
beautifulsoup4
newspaper3k
lxml_html_clean
